# build/swagger/swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"IUserModel": {
			"properties": {
				"id": {
					"type": "string"
				},
				"username": {
					"type": "string"
				},
				"firstName": {
					"type": "string"
				},
				"lastName": {
					"type": "string"
				}
			},
			"required": [
				"username",
				"firstName",
				"lastName"
			],
			"type": "object"
		},
		"PaginationModel": {
			"properties": {
				"count": {
					"type": "number",
					"format": "double"
				},
				"pageNumber": {
					"type": "number",
					"format": "double"
				},
				"perPage": {
					"type": "number",
					"format": "double"
				},
				"list": {
					"type": "array",
					"items": {
						"type": "object"
					}
				},
				"args": {
					"$ref": "#/definitions/PaginationModel"
				}
			},
			"required": [
				"count",
				"pageNumber",
				"perPage",
				"list",
				"args"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "express-swagger-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/user/{id}": {
			"get": {
				"operationId": "GetById",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IUserModel"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"security": [
					{
						"adminUser": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "userParams",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUserModel"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"adminUser": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/user": {
			"get": {
				"operationId": "GetPaginated",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginationModel"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "pageNumber",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "perPage",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IUserModel"
						}
					}
				},
				"security": [
					{
						"adminUser": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "userParams",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUserModel"
						}
					}
				]
			}
		},
		"/ping": {
			"get": {
				"operationId": "Ping",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}